feat: create CLI for generation.
BREAKING CHANGE: since the dart build
system is not as dynamic as it should be for
this usecase, using a builder is not
sufficient. Now a CLI executable
is used to find and parse the design system
configuration and then generate the
specific design token files according to
the configuration.

Signed-off-by: Christoph Bühler <christoph@smartive.ch>